backend/requirements.txt
absl-py
astor
cachetools
pandas
certifi
chardet
Click
cycler
Flask
xgboost
pickle5
gast
google-api-core
google-auth
google-cloud-core
google-cloud-storage
google-resumable-media
googleapis-common-protos
grpcio
gunicorn
h5py
idna
itsdangerous
Jinja2
kiwisolver
Markdown
MarkupSafe
Flask-Cors
scikit-learn
matplotlib
mock
numpy
praat-parselmouth
protobuf
pyasn1
pyasn1-modules
pyparsing
python-dateutil
pytz
requests
rsa
scipy
six
termcolor
urllib3
Werkzeug
